problem_4/requirements.txt
pkg_resources==0.0.0
psycopg2-binary==2.9.5
